Refactoring mcverry_2006
